rename to test-es6
