--- docs/tests/v3/es6/css/sanddance.css
@@ -1,28 +1,0 @@
-.sanddance-gl {
-    border: 1px solid #ccc;
-    float: left;
-    height: 700px;
-    width: 700px;
-    margin-right: 1em;
-}
-
-.sanddance-tooltip table {
-    background: #333;
-    color: #fff;
-    font-size: smaller;
-    margin: 1em;
-    min-width: 16em;
-    padding: 6px;
-    position: absolute; 
-}
-  
-.sanddance-tooltip td {
-    text-align: left;
-    vertical-align: top;
-    width: 75%; 
-}
-
-.sanddance-tooltip td:first-child {
-    width: 25%; 
-}
-
