--- docs/tests/v3/es6/js/sanddance.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>sanddance test</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="css/sanddance.css" />
-</head>
-
-<body>
-    <header>
-        <h1>sanddance test</h1>
-    </header>
-
-    <div id="vis"></div>
-
-    <script src="js/sanddance.js"></script>
-
-</body>
-
-</html>
